# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayItem API",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/token": {
      "post": {
        "summary": "Get access token (demo)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string"
                  }
                },
                "required": [
                  "client_id",
                  "client_secret"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token"
          }
        }
      }
    },
    "/api/payitem": {
      "post": {
        "summary": "Submit payitem",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ClientID": {
                    "type": "string"
                  },
                  "SchemeID": {
                    "type": "string"
                  },
                  "EmployeeID": {
                    "type": "string"
                  },
                  "Payitemname": {
                    "type": "string"
                  },
                  "Value": {
                    "type": "number"
                  },
                  "Effectivedate": {
                    "type": "string",
                    "pattern": "^\\d{2}/\\d{2}/\\d{4}$",
                    "example": "07/08/2025"
                  }
                },
                "required": [
                  "ClientID",
                  "SchemeID",
                  "EmployeeID",
                  "Payitemname",
                  "Value",
                  "Effectivedate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "summary": "List payitems (filtered by date range)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "YYYY-MM-DD"
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "YYYY-MM-DD"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}